+++ activesupport/lib/active_support/core_ext/string/inflections.rb
@@ -17,7 +17,7 @@ class String
   #
   # If the optional parameter +locale+ is specified,
   # the word will be pluralized as a word of that language.
-  # By default, this parameter is set to <tt>:en</tt>.
+  # By default, this parameter is set to the current I18n locale.
   # You must define your own inflection rules for languages other than English.
   #
   #   'post'.pluralize             # => "posts"
@@ -30,8 +30,8 @@ class String
   #   'apple'.pluralize(2)         # => "apples"
   #   'ley'.pluralize(:es)         # => "leyes"
   #   'ley'.pluralize(1, :es)      # => "ley"
-  def pluralize(count = nil, locale = :en)
-    locale = count if count.is_a?(Symbol)
+  def pluralize(count = nil, locale = nil)
+    locale = count.is_a?(Symbol) ? count : (locale || I18n.locale || :en)
     if count == 1
       dup
     else
@@ -43,7 +43,7 @@ class String
   #
   # If the optional parameter +locale+ is specified,
   # the word will be singularized as a word of that language.
-  # By default, this parameter is set to <tt>:en</tt>.
+  # By default, this parameter is set to the current I18n locale.
   # You must define your own inflection rules for languages other than English.
   #
   #   'posts'.singularize            # => "post"
@@ -53,7 +53,8 @@ class String
   #   'the blue mailmen'.singularize # => "the blue mailman"
   #   'CamelOctopi'.singularize      # => "CamelOctopus"
   #   'leyes'.singularize(:es)       # => "ley"
-  def singularize(locale = :en)
+  def singularize(locale = nil)
+    locale = locale || I18n.locale || :en
     ActiveSupport::Inflector.singularize(self, locale)
   end